# bench: Make WalletLoading benchmark run faster