--- a/src/bench/wallet_loading.cpp
+++ b/src/bench/wallet_loading.cpp
@@ -68,7 +68,7 @@ static void WalletLoading(benchmark::Bench& bench, bool legacy_wallet)
     // reload the wallet for the actual benchmark
     BenchUnloadWallet(std::move(wallet));
 
-    bench.minEpochIterations(10).run([&] {
+    bench.run([&] {
         wallet = BenchLoadWallet(context, options);
 
         // Cleanup