`minEpochIterations` is probably unnecessary to set, so removing it makes the runtime much faster.