@w0xlt 2022-04-19
On my machine:

Before PR #24913
```
Executed in   57.34 secs    fish           external
   usr time   57.60 secs  843.00 micros   57.60 secs
   sys time    1.43 secs    0.00 micros    1.43 secs
```
After PR #24913
```
|    1,525,136,424.90 |                0.66 |    0.2% |    189.67 | `WalletLoadingDescriptors`
|      873,902,969.45 |                1.14 |    0.3% |    106.20 | `WalletLoadingLegacy`

________________________________________________________
Executed in  564.01 secs    fish           external
   usr time  540.03 secs  427.00 micros  540.02 secs
   sys time   13.35 secs   95.00 micros   13.35 secs
```
This PR
```
|    1,740,638,493.00 |                0.57 |    0.6% |     26.13 | `WalletLoadingDescriptors`
|    1,016,725,382.00 |                0.98 |    0.9% |     12.72 | `WalletLoadingLegacy`

________________________________________________________
Executed in  348.35 secs    fish           external
   usr time  328.66 secs    1.05 millis  328.66 secs
   sys time   12.46 secs    0.60 millis   12.46 secs
```

There's a good improvement  (348.35 secs) over PR #24913 (540.03 secs), but it still takes a lot longer than before it (57.34 secs).