@fanquake 2022-04-20
> @fanquake Is this good enough or do you want faster?

I think to be good enough, it needs to run in a similar time to all of the other benchmarks, i.e max ~1-2 seconds, otherwise anyone running `make check` on macOS is going to be opening issues asking why running `make check` (has suddenly started) taking 10+ minutes. I tested this branch, but stopped after `WalletLoadingDescriptors` took 12 minutes:
```bash
|           80,237.38 |           12,463.02 |    0.1% |      0.01 | `VerifyScriptBench`
|            8,190.91 |          122,086.52 |    1.2% |      0.01 | `WalletBalanceClean`
|          179,780.20 |            5,562.35 |    1.5% |      0.01 | `WalletBalanceDirty`
|            7,929.42 |          126,112.61 |    0.1% |      0.01 | `WalletBalanceMine`
|               42.47 |       23,546,683.62 |    0.0% |      0.01 | `WalletBalanceWatch`
|      902,283,977.00 |                1.11 |    1.4% |    105.48 | `WalletLoadingDescriptors`
^C
src/bench/bench_bitcoin  59.66s user 24.31s system 11% cpu 12:01.03 total
```

> -unsafesqlitesync does not appear to have any effect. The next thing would be to reduce the number of txs created, but I'm concerned that doing so will make it harder to observe actual performance issues that this benchmark is intended to measure.

Given that, it seems like we've got another macOS only-performance issue, that we should probably figure out the root cause of?