@laanwj 2022-04-20
> I think to be good enough, it needs to run in a similar time to all of the other benchmarks, i.e max ~1-2 seconds, 

Agree. It's unacceptable to have a benchmark take in the order of minutes. Remember that benchmarks are there to measure small timespans (such as inner loop functions), not very long processes.

I've also seen a CI timeout due to this. Remember that we do `bench/bench_bitcoin &> /dev/null` in one of the CI runs.