@achow101 2022-04-20
I've added the following changes:
* The number of benchmark epochs is reduced from the default 11 to 5.
* Use `unsafesqlitesync`
* Use mock databases
  * Update sqlite's mock database implementation to use in-memory dbs rather than just temp dbs
* Instead of mining blocks to wallet addresses, we make transactions and add them directly with `AddToWallet`.
* Reduced the number of txs added to the wallet from 5000 to 1000.

This still doesn't make it as fast as the other benchmarks though.

***

Before #24913:

```
________________________________________________________
Executed in   11.37 secs    fish           external
   usr time   11.36 secs    0.00 micros   11.36 secs
   sys time    7.62 secs  435.00 micros    7.62 secs
```

After #24913:

```
________________________________________________________
Executed in  150.77 secs    fish           external
   usr time  148.99 secs  334.00 micros  148.99 secs
   sys time    9.42 secs   88.00 micros    9.42 secs
```

This PR:

```
________________________________________________________
Executed in   20.25 secs    fish           external
   usr time   20.21 secs  319.00 micros   20.21 secs
   sys time    7.92 secs  158.00 micros    7.92 secs
```