@sipa 2022-05-10
This PR makes `./src/bench/bench_bitcoin -filter=WalletLoadingDescriptors` go from taking 246s to 2.3s for me, making "make check" a lot more usable.

(Ubuntu 22.04 on a Ryzen 5950X system; my ZFS filesystem setup may make flushing/syncing relatively slower)