@Rspigler 2022-05-22
tACK e673d8b475995075b696208386c9e45ae7ca3e20

This solves the `make check` issue with `bench_bitcoin` reported [here](https://github.com/bitcoin/bitcoin/issues/25185)