@achow101 2022-05-26
> if I (manually) revert commits [817c051](https://github.com/bitcoin/bitcoin/commit/817c051364208d3f9e7e2af5700bd2bee5c9f303) (using unsafesqlitesync)

I think `-unsafesqlitesync` has more of an effect on macs than linux?